InterfaceManager class updated
